Synergies Prototype 2
Fixed up a bit for the prototypes for multiple, and also turns out the Console was wrongly called Block Coding the entire time.  Added 6 new updates for each generator.  Also, BTW, the save file is not blanck because it's easier for testing than if you had to start all over again.