GameEngine Prototype 2
Checked 1 more time, everything seems to be good.  not many changes really, just a final check and 3 more GameEngine upgrades.  Fixed 1 bug which was that it was trying to get the autoclicker cooldown even though there was not a autoclicker. Going to release next version on Saturday. some more small fixes. was going to commit then realized om some other fmall bugs like it siwtching to per second then back which it confusing so i just made it not go to the per second, becasue its just not implemented yet or working.